move mocks to mocks package
