+++ mocks/mock_api.go
@@ -1,0 +1,9 @@
+package mocks
+
+type mockApi struct {
+    handleRequestsCalls uint64
+}
+
+func (m *mockApi) HandleRequests() {
+    m.handleRequestsCalls++
+}
