+++ mocks/mock_verifier.go
@@ -1,0 +1,9 @@
+package mocks
+
+type mockVerifier struct {
+    verifyCalls uint64
+}
+
+func (m *mockVerifier) Verify() {
+    m.verifyCalls++
+}
