+++ reader/test_helpers.go
@@ -5,23 +5,6 @@ import (
 	"os"
 )
 
-type mockQueue struct {
-	PutCalls uint64
-}
-
-func (m mockQueue) Size() int {
-	return -1
-}
-
-func (m mockQueue) Get() (string, error) {
-	return "", nil
-}
-
-func (m *mockQueue) Put(password string) error {
-	m.PutCalls++
-	return nil
-}
-
 func setupFile(testPath string, lines []string) {
 	file, _ := os.Create(testPath)
 	defer file.Close()
